fix: resolve Next.js route type error by extracting utility function
