+++ lib/flashcard-submit-utils.ts
@@ -1,0 +1,29 @@
+/*
+ * SakuMari - Japanese Kana Flashcard App
+ * Copyright (C) 2025  Sakan Nirattisaykul
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+import { NextResponse } from "next/server";
+import { ApiErrors } from "@/lib/api-errors";
+
+export function handleSubmissionError(error: unknown): NextResponse {
+  if (error instanceof Error) {
+    if (error.message.includes("Invalid JSON")) {
+      return ApiErrors.badRequest("Invalid request format");
+    }
+  }
+
+  return ApiErrors.internalError("Failed to submit flashcard answer");
+}
